# backend/src/docs/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Painting shop administration API
  description: Painting shop administration API
servers:
  - url: http://localhost:3000
paths:
...
  /photo:
    get:
      description: Returns a list of photo
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of photo returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Photo'
        '501':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      description: Creates a new photo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Photo'
      responses:
        '201':
          description: Photo has been created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '501':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /photo/{_id}:
    get:
      description: Get an photo
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: _id
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Returns an photo
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/Photo'
        '501':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      description: Updates az photo
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: _id
          required: true
          schema: 
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Photo'
      responses:
        '200':
          description: Photo has been updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '501':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      description: Removes an photo
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: _id
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Photo has been deleted successfully
          content:
            application/json:
              schema:
                type: object
                example: {}
        '501':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Photo:
      type: object
      required:
        - storedFileName
      properties:
        storedFileName:
          type: string
        alt:
          type: string
        fileSize:
          type: number
...
    Error:
      type: object
      required:
        - message
        - error
      properties:
        message:
          type: string
        error:
          type: string